+++ Lib/fontbakery/checks/universal/__init__.py
@@ -2483,18 +2483,32 @@ def com_google_fonts_check_tabular_kerning(ttFont):
         hb.shape(vhb.hbfont, buf, features)
         return buf
 
-    def kerning(ttFont, glyph_list):
+    def get_kerning(ttFont, glyph_list):
         GID_list = [GID_for_glyph(ttFont, glyph) for glyph in glyph_list]
-        return buf_to_width(
+        width1 = buf_to_width(
             buffer_for_GIDs(
                 GID_list,
                 {"kern": True},
             )
-        ) - buf_to_width(
+        )
+        width2 = buf_to_width(
             buffer_for_GIDs(
                 GID_list,
                 {"kern": False},
             )
+        )
+        return width1 - width2
+
+    def get_str_buffer(ttFont, glyph_list):
+        GID_list = [GID_for_glyph(ttFont, glyph) for glyph in glyph_list]
+        buffer = buffer_for_GIDs(GID_list, {})
+        string = vhb.serialize_buf(buffer)
+        return string
+
+    def digraph_kerning(ttFont, glyph_list, expected_kerning):
+        return (
+            len(get_str_buffer(ttFont, glyph_list).split("|")) > 1
+            and get_kerning(ttFont, glyph_list) == expected_kerning
         )
 
     all_glyphs = list(ttFont.getGlyphOrder())
@@ -2545,19 +2559,24 @@ def com_google_fonts_check_tabular_kerning(ttFont):
             (tabular_numerals, tabular_glyphs),
         ):
             combinations = unique_combinations(sets[0], sets[1])
-            for a, b in combinations:
-                if kerning(ttFont, [a, b]) != 0:
-                    yield FAIL, Message(
-                        "has-tabular-kerning",
-                        f"Kerning between {a} and {b} is not zero",
-                    )
-                    passed = False
-                if kerning(ttFont, [b, a]) != 0:
-                    yield FAIL, Message(
-                        "has-tabular-kerning",
-                        f"Kerning between {b} and {a} is not zero",
-                    )
-                    passed = False
+            for x, y in combinations:
+                for a, b in ((x, y), (y, x)):
+                    kerning = get_kerning(ttFont, [a, b])
+                    if kerning != 0:
+                        # Check if either a or b are digraphs that themselves
+                        # have kerning when decomposed in ccmp
+                        # that would throw off the reading, skip if it's identical
+                        # to the kerning of the whole sequence
+                        if digraph_kerning(ttFont, [a], kerning) or digraph_kerning(
+                            ttFont, [b], kerning
+                        ):
+                            pass
+                        else:
+                            yield FAIL, Message(
+                                "has-tabular-kerning",
+                                f"Kerning between {a} and {b} is {kerning}, should be 0",
+                            )
+                            passed = False
 
     if passed:
         yield PASS, "No kerning found for tabular glyphs"